requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-auth-ldap==2.4.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
greenlet==1.1.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mod-wsgi==4.9.0
mssql-django==1.0rc1
numpy==1.21.0
packaging==21.0
pandas==1.3.0
pandasql==0.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyodbc==4.0.31
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-ldap @ file:///E:/devs/python_ldap-3.3.1-cp39-cp39-win_amd64.whl
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
